Change animation settings source to attribute.
